requirements.txt
pdfkit==1.0.0
pypdf==3.2.0
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.4.0
